requirements.txt
requests
lxml
bs4
pandas
numpy
scrapy
selenium
webdriver_manager
python-dotenv
langchain
tqdm